the cost handling
Thte cost is calculated by the minutes that the users book a supporter. After finishing the job, the supporter can get the money and add to the credit point. IF the member go to the comment part when finishing the job, the deducting in money will be occured in users bank account